add level change request

+++ server/src/main/java/at/ac/tuwien/foop/server/game/Configuration.java
@@ -1,5 +1,5 @@
 package at.ac.tuwien.foop.server.game;
 
 
-public record Configuration(int tickRate, int mouseSpeed, int catSpeed) {
+public record Configuration(int tickRate, double mouseSpeed, double catSpeed) {
 }
